+++ keyring/cli.py
@@ -7,8 +7,9 @@ import getpass
 from optparse import OptionParser
 import sys
 
-from . import get_keyring, set_keyring, get_password, set_password, delete_password
 from . import core
+from . import backend
+from . import get_keyring, set_keyring, get_password, set_password, delete_password
 
 
 class CommandLineTool(object):
@@ -21,9 +22,17 @@ class CommandLineTool(object):
         self.parser.add_option("-b", "--keyring-backend",
                                dest="keyring_backend", default=None,
                                help="Name of the keyring backend")
+        self.parser.add_option("--list-backends",
+                               action="store_true",
+                               help="List keyring backends and exit")
 
     def run(self, argv):
         opts, args = self.parser.parse_args(argv)
+
+        if opts.list_backends:
+            for k in backend.get_all_keyring():
+                print(k)
+            return
 
         try:
             kind, service, username = args
@@ -39,8 +48,7 @@ class CommandLineTool(object):
             try:
                 if opts.keyring_path:
                     sys.path.insert(0, opts.keyring_path)
-                backend = core.load_keyring(opts.keyring_backend)
-                set_keyring(backend)
+                set_keyring(core.load_keyring(opts.keyring_backend))
             except (Exception,):
                 # Tons of things can go wrong here:
                 #   ImportError when using "fjkljfljkl"