+++ CHANGES.rst
@@ -1,6 +1,9 @@
 10.5.0
 ------
 
+* #289: Added ``--list-backends`` option to
+  command-line interface.
+
 * Removed ``logger`` from ``keyring``. See #291 for related
   request.